backend/requirements.txt
django[bcrypt]
psycopg2
django-celery-email
django-redis
uwsgi
djangorestframework
django-filter
drf-extensions
django-ipware
django-extensions
bpython
pygraphviz
werkzeug
watchdog
django-debug-toolbar
ptvsd==3.0.0
